Move TestData into test init
